testinggg

+++ DTOs/Dashboard/WeeklySalesReportDTO.cs
@@ -6,8 +6,7 @@
         public DateTime WeekStartDate { get; set; }
         public DateTime WeekEndDate { get; set; }
         public int NumberOfOrders { get; set; } // Count should be an integer
-        public decimal AverageOrderValue { get; set; } // Average should be a decimal
-        public decimal TotalSalesAmount { get; set; } // Total sales amount should be decimal
+        public double AverageOrderValue { get; set; } // Average as double
+        public double TotalSalesAmount { get; set; } // Total sales amount as double
     }
-
 }
